generate coverage reports

+++ app/trainers/medcat_deid_trainer.py
@@ -180,7 +180,7 @@ class MedcatDeIdentificationSupervisedTrainer(MedcatSupervisedTrainer):
                     trainer._training_in_progress = False
                 trainer._housekeep_file(model_pack_path)
                 trainer._housekeep_file(copied_model_pack_path)
-                if cdb_config_path:
+                if cdb_config_path and os.path.exists(cdb_config_path):
                     os.remove(cdb_config_path)
         else:
             try:
