+++ app/trainers/metacat_trainer.py
@@ -116,7 +116,7 @@ class MetacatTrainer(MedcatSupervisedTrainer):
                     trainer._training_in_progress = False
                 trainer._housekeep_file(model_pack_path)
                 trainer._housekeep_file(copied_model_pack_path)
-                if cdb_config_path:
+                if cdb_config_path and os.path.exists(cdb_config_path):
                     os.remove(cdb_config_path)
         else:
             try:
